requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
gradio==5.43.1
requests==2.32.5
pydantic==2.11.7
numpy==2.3.2
pandas==2.3.1
scikit-learn==1.7.1
scipy==1.16.1
joblib==1.5.1
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
python-dotenv==1.1.1
xgboost==3.0.4
pytest==8.4.1
pytest-cov==6.2.1
gunicorn

